+++ src/cow-react/utils/amountFormat/index.test.ts
@@ -1,7 +1,7 @@
 import { formatAmountWithPrecision, formatFiatAmount, formatPercent, formatTokenAmount } from './index'
 import { CurrencyAmount, Percent } from '@uniswap/sdk-core'
 import { DAI_GOERLI } from 'utils/goerli/constants'
-import { USDC_GNOSIS_CHAIN } from '../../../custom/utils/gnosis_chain/constants'
+import { USDC_GNOSIS_CHAIN } from '../../../utils/gnosis_chain/constants'
 
 describe('Amounts formatting', () => {
   const decimals = DAI_GOERLI.decimals
