+++ src/custom/hooks/useGetReceipt.ts
@@ -3,7 +3,7 @@ import { useWeb3React } from '@web3-react/core'
 import { retry, RetryableError, RetryOptions } from 'utils/retry'
 import { TransactionReceipt } from '@ethersproject/abstract-provider'
 import { RetryResult } from 'types'
-import { supportedChainId } from '../utils/supportedChainId'
+import { supportedChainId } from 'utils/supportedChainId'
 import { useWalletInfo } from '@cow/modules/wallet'
 
 const DEFAULT_RETRY_OPTIONS: RetryOptions = { n: 3, minWait: 1000, maxWait: 3000 }
