+++ src/custom/hooks/useWrapCallback.ts
@@ -11,7 +11,7 @@ import { RADIX_HEX } from '@src/constants'
 import { formatSymbol } from '@cow/utils/format'
 import { getOperationMessage, OperationType } from 'components/TransactionConfirmationModal'
 import { calculateGasMargin } from 'utils/calculateGasMargin'
-import { isRejectRequestProviderError } from '../utils/misc'
+import { isRejectRequestProviderError } from '../../utils/misc'
 import { wrapAnalytics } from 'components/analytics'
 import { useDerivedSwapInfo } from 'state/swap/hooks'
 import { useCloseModals } from 'state/application/hooks'
