+++ src/lib/hooks/useCurrency.ts
@@ -8,9 +8,9 @@ import useNativeCurrency from 'lib/hooks/useNativeCurrency'
 import { useMemo } from 'react'
 import { isChainAllowed } from '@cow/modules/wallet/web3-react/connection'
 
-import { TOKEN_SHORTHANDS } from '../../constants/tokens'
-import { isAddress } from '../../utils'
-import { supportedChainId } from '../../utils/supportedChainId'
+import { TOKEN_SHORTHANDS } from 'constants/tokens'
+import { isAddress } from 'utils'
+import { supportedChainId } from 'utils/supportedChainId'
 import { useWalletInfo } from '@cow/modules/wallet'
 
 // parse a name or symbol from a token response
