+++ src/utils/gnosis_chain/constants.ts
