+++ src/utils/gnosis_chain/hack.ts
